Introduced NutracWsgiApplication to the project.
We need a data connected object so we can attach services to the
wsgi application process.

Before we had only a function called application, this function
would return the Trac application if everything was correct.

The process method will return the trac application and we can
attach services to it.

Refs: #1